Quote_Comparison_Chatbot/requirements.txt
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
flask
sentence-transformers